@@ -10,8 +10,8 @@
     "@mui/material": "^5.10.2",
     "@tanstack/react-query": "^4.2.3",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
@@ -61,7 +61,7 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.4.3",
     "eslint-config-airbnb": "^19.0.4"
   }
